extract some interfaces
