+++ Homework/TagsCloud.Visualization/IWordsService.cs
@@ -1,0 +1,9 @@
+﻿using TagsCloud.Visualization.Models;
+
+namespace TagsCloud.Visualization
+{
+    public interface IWordsService
+    {
+        Word[] GetWords();
+    }
+}
